glfw/requirements.txt
PyOpenGL
PyOpenGL_accelerate
numpy
glfw
PyGLM